Added the GCC error message format to reporters and tidy those.
