+++ packages/markuplint/src/reporter/index.ts
@@ -1,1 +1,4 @@
-export * from './reporter';
+export * from './standard-reporter';
+export * from './simple-reporter';
+export * from './gcc-reporter';
+export * from './types';
